Split project structure
