+++ .gitignore
@@ -1,4 +1,4 @@
-nesthub
+/nesthub
 config.json
 oauth.json
 *.swp
