+++ internal/helpers/openurl.go
@@ -1,11 +1,11 @@
 //go:build !darwin
 // +build !darwin
 
-package main
+package helpers
 
 import "fmt"
 
-func openURL(u string) error {
+func OpenURL(u string) error {
 	fmt.Println("Please open", u, "in browser")
 	return nil
 }
